Avoid 32-bit divisions
This change reduces running time of mnist/baseline by 17%

+++ common/intermittent-cnn.h
@@ -10,7 +10,7 @@ struct ParameterInfo;
 struct Model;
 uint8_t run_cnn_tests(uint16_t n_samples);
 
-uint32_t job_index_to_offset(const ParameterInfo* output, uint32_t job_index);
+uint32_t job_index_to_offset(const ParameterInfo* output, uint16_t job_index);
 
 uint8_t get_state_bit(struct Model *model, uint8_t slot_id);
 #if STATEFUL
